# [Core] in batch prefix caching by delay scheduling