<!-- Thank you for your contribution! We appreciate it. The following guidelines will help improve your pull request and facilitate feedback. If anything is unclear, don't hesitate to submit your pull request and ask the maintainers for assistance. -->

## Motivation

Implement in-batch prefix caching. 

If a new request has a short enough matching prefix, and if there are more than 1 such request that share the same prefix, we deprioritize them except one of them.

The feature benefits the workloads with many shared prefix prompts. Here's the extreme example we can be benefited from this feature. 


```
    prompts = [
        *["Hello, my name is " * 50 + chr(i) for i in range(65, 87)],
        *[
            "The president of the United States is " * 50 + chr(i)
            for i in range(65, 87)
        ],
        *["capital of France is " * 50 + chr(i) for i in range(65, 87)],
        *["The future of AI is " * 50 + chr(i) for i in range(65, 87)],
        *["dkfjklwerkj lskjdfsa " * 50 + chr(i) for i in range(65, 87)],
        *["What time is it now? " * 50 + chr(i) for i in range(65, 87)],
        *["hello it's good to see you " * 50 + chr(i) for i in range(65, 87)],
        *["xai is the " * 50 + chr(i) for i in range(65, 87)],
        *["oops " * 50 + chr(i) for i in range(65, 87)],
        *["hello it's me " * 50 + chr(i) for i in range(65, 87)],
    ]


```

With this prompt, we can see 2X+ perf improvement

```
 after
e2e takes 807.3112380225211 ms
# before
e2e takes 1828.76543502789 ms
```

I also added `bench_prefix.py` script which is originally written from @Ying1123. With prefix_len=1024, gen_len=128, and sampling_size = 32, I could see 10+% performance improvement

```
# lpm-2
Throughput: 94.79 requests/s, 109553.50 tokens/s
# lpm
Throughput: 83.51 requests/s, 96514.85 tokens/s
```

The feature adds a slight more overhead to `calc_priroity` because it requires additional radix cache operation. I believe it will be negligible with overlap schedule. i also verified there's no regression from `python -m sglang.bench_offline_throughput --model-path meta-llama/Meta-Llama-3.1-8B-Instruct --num-prompts 10` just in case. 

## Modifications

<!-- Describe the changes made in this PR. -->

## Checklist

- [ ] Format your code according to the [Contributor Guide](https://github.com/sgl-project/sglang/blob/main/docs/references/contributor_guide.md).
- [ ] Add unit tests as outlined in the [Contributor Guide](https://github.com/sgl-project/sglang/blob/main/docs/references/contributor_guide.md).
- [ ] Update documentation as needed, including docstrings or example tutorials.