@ByronHsu 2024-12-11
Awesome enhancement! Can we use https://github.com/sgl-project/sglang/pull/1990 instead of adding a new bench script?