@ByronHsu 2024-12-11
cc @MrAta i think this is useful for our case too