@rkooo567 2024-12-11
- removed benchmark_prefix.py (use https://github.com/sgl-project/sglang/pull/1990 instead)
- Support dfs-weight. For the same extreme prompts, I see 2x improvement

```
# dfs-weight after
e2e takes 778.6834560101852 ms
# dfs-weight before
e2e takes 1497.2664599772543 ms

```